Fix warning: implicit conversion changes signedness
visp/modules/tracker/dnn/src/vpMegaPose.cpp: In function ‘void encode(std::vector<unsigned char>&, const T&) [with T = int]’:
visp/modules/tracker/dnn/src/vpMegaPose.cpp:92:28: warning: conversion to ‘__uint32_t’ {aka ‘unsigned int’} from ‘int’ may change the sign of the result [-Wsign-conversion]
   92 |   const uint32_t v = htonl(object);
      |                            ^~~~~~

+++ modules/tracker/dnn/src/vpMegaPose.cpp
@@ -89,7 +89,7 @@ void encode(std::vector<uint8_t> &buffer, const T &object) = delete;
 template<>
 void encode(std::vector<uint8_t> &buffer, const int &object)
 {
-  const uint32_t v = htonl(object);
+  const uint32_t v = htonl(static_cast<uint32_t>(object));
   const uint8_t *varr = (uint8_t *)&v;
   buffer.insert(buffer.end(), varr, varr + 4);
 }
